Fixed the featureVectorSize checker in umpackMFC.py, which was not catching the errors it was supposed to chatch. It was supposed to catch inconsistencies between the given featureVectorSize and the actual feature vector size by checking whether the total size of the MFCC file indicated the given featureVectorSize was even possible.

+++ unpackMFC.py
@@ -8,7 +8,7 @@ def run(input, featureVectorSize):
 	sign = 1;
 	file = open(input, 'r')
 	size = struct.unpack('>i', ''.join(file.read(4)))[0]
-	if size / featureVectorSize - (float)(size // featureVectorSize) != 0:
+	if ((float)(size)) / featureVectorSize - (float)(size // featureVectorSize) != 0:
 		print "ERR: unpackMFC.run().featureVectorSize is inconsistent with the feature count read from the file given."
 		print "File given: ", input
 		print "Feature count read: ", size
